src/requirements.api.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tensorflow-cpu==2.17.0
pillow==10.1.0
python-multipart==0.0.6
numpy<2.0.0
tf-keras